requirements.txt
tensorflow>=2.10.0
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.23.0
Pillow>=9.0.0
matplotlib>=3.6.0
scikit-learn>=1.1.0
werkzeug>=2.3.0
